Fix broken webpage caused by overflowing table
This commit fixes a mobile/small screen responsiveness issue on the tRPC webpage ("Awesome tRPC Collection" webpage under "Community") by adding 100% width and table-layout: fixed to table element. 

Previously, the content of the tables was causing an overflow and breaking the webpage on mobile devices. By setting the table width to 100% and using a fixed layout algorithm, the tables now adjust their size to fit the width of the parent container and maintain consistent column widths.

This ensures that the webpage is responsive on all mobile/small screen devices and improves the overall user experience.